feat: Add act to unregister identity
